fix: renamed getCityVehicleAvailabilityTextColor to getVehicleAvailabilityTextColor
