+++ app/component/StopsNearYouContainer.js
@@ -7,7 +7,7 @@ import { matchShape } from 'found';
 import StopNearYouContainer from './StopNearYouContainer';
 import withBreakpoint from '../util/withBreakpoint';
 import { sortNearbyRentalStations, sortNearbyStops } from '../util/sortUtils';
-import CityBikeStopNearYou from './CityVehicleStopNearYou';
+import CityBikeStopNearYou from './VehicleStopNearYou';
 import Loading from './Loading';
 import Icon from './Icon';
 import { getDefaultNetworks } from '../util/vehicleRentalUtils';
